Adding cpp and python formatting changes

+++ rocAL/rocAL_pybind/amd/rocal/plugin/tf.py
@@ -23,6 +23,7 @@ import cupy as cp
 import rocal_pybind as b
 import amd.rocal.types as types
 
+
 class ROCALGenericImageIterator(object):
     def __init__(self, pipeline):
         self.loader = pipeline
@@ -43,7 +44,7 @@ class ROCALGenericImageIterator(object):
             for i in range(len(self.output_tensor_list)):
                 self.dimensions = self.output_tensor_list[i].dimensions()
                 self.dtype = self.output_tensor_list[i].dtype()
-                self.output = np.empty(self.dimensions, dtype = self.dtype)
+                self.output = np.empty(self.dimensions, dtype=self.dtype)
 
                 self.output_tensor_list[i].copy_data(self.output)
                 self.output_list.append(self.output)
@@ -60,6 +61,7 @@ class ROCALGenericImageIterator(object):
 
     def __del__(self):
         b.rocalRelease(self.loader._handle)
+
 
 class ROCALGenericIteratorDetection(object):
     def __init__(self, pipeline, tensor_layout=types.NCHW, reverse_channels=False, multiplier=None, offset=None, tensor_dtype=types.FLOAT, device=None, device_id=0):
@@ -82,10 +84,10 @@ class ROCALGenericIteratorDetection(object):
     def __next__(self):
         if self.loader.rocal_run() != 0:
             timing_info = self.loader.timing_info()
-            print("Load     time ::",timing_info.load_time)
-            print("Decode   time ::",timing_info.decode_time)
-            print("Process  time ::",timing_info.process_time)
-            print("Transfer time ::",timing_info.transfer_time)
+            print("Load     time ::", timing_info.load_time)
+            print("Decode   time ::", timing_info.decode_time)
+            print("Process  time ::", timing_info.process_time)
+            print("Transfer time ::", timing_info.transfer_time)
             raise StopIteration
         self.output_tensor_list = self.loader.get_output_tensors()
 
@@ -100,14 +102,15 @@ class ROCALGenericIteratorDetection(object):
                     self.output_tensor_list[i].copy_data(self.output)
                 else:
                     self.output = cp.empty(self.dimensions, dtype=self.dtype)
-                    self.output_tensor_list[i].copy_data(self.output.data.ptr) 
+                    self.output_tensor_list[i].copy_data(self.output.data.ptr)
                 self.output_list.append(self.output)
         else:
             for i in range(len(self.output_tensor_list)):
                 if self.device == "cpu":
                     self.output_tensor_list[i].copy_data(self.output_list[i])
                 else:
-                    self.output_tensor_list[i].copy_data(self.output_list[i].data.ptr)
+                    self.output_tensor_list[i].copy_data(
+                        self.output_list[i].data.ptr)
 
         if self.loader._name == "TFRecordReaderDetection":
             self.bbox_list = []
@@ -121,7 +124,8 @@ class ROCALGenericIteratorDetection(object):
             self.num_bboxes_list = []
             self.loader.get_img_sizes(self.img_size)
             for i in range(self.bs):
-                self.label_2d_numpy = np.reshape(self.labels[i], (-1, 1)).tolist()
+                self.label_2d_numpy = np.reshape(
+                    self.labels[i], (-1, 1)).tolist()
                 self.bb_2d_numpy = np.reshape(self.bboxes[i], (-1, 4)).tolist()
                 self.num_bboxes_list.append(len(self.bboxes[i]))
                 self.label_list.append(self.label_2d_numpy)
@@ -129,18 +133,24 @@ class ROCALGenericIteratorDetection(object):
 
             self.target = self.bbox_list
             self.target1 = self.label_list
-            max_cols = max([len(row) for batch in self.target for row in batch])
+            max_cols = max([len(row)
+                           for batch in self.target for row in batch])
             # max_rows = max([len(batch) for batch in self.target])
             max_rows = 100
-            bb_padded = [batch + [[0] * (max_cols)] * (max_rows - len(batch)) for batch in self.target]
-            bb_padded_1 = [row + [0] * (max_cols - len(row)) for batch in bb_padded for row in batch]
+            bb_padded = [
+                batch + [[0] * (max_cols)] * (max_rows - len(batch)) for batch in self.target]
+            bb_padded_1 = [row + [0] * (max_cols - len(row))
+                           for batch in bb_padded for row in batch]
             arr = np.asarray(bb_padded_1)
             self.res = np.reshape(arr, (-1, max_rows, max_cols))
-            max_cols = max([len(row) for batch in self.target1 for row in batch])
+            max_cols = max([len(row)
+                           for batch in self.target1 for row in batch])
             # max_rows = max([len(batch) for batch in self.target1])
             max_rows = 100
-            lab_padded = [batch + [[0] * (max_cols)] * (max_rows - len(batch)) for batch in self.target1]
-            lab_padded_1 = [row + [0] * (max_cols - len(row)) for batch in lab_padded for row in batch]
+            lab_padded = [
+                batch + [[0] * (max_cols)] * (max_rows - len(batch)) for batch in self.target1]
+            lab_padded_1 = [row + [0] * (max_cols - len(row))
+                            for batch in lab_padded for row in batch]
             labarr = np.asarray(lab_padded_1)
             self.l = np.reshape(labarr, (-1, max_rows, max_cols))
             self.num_bboxes_arr = np.array(self.num_bboxes_list)
@@ -149,13 +159,19 @@ class ROCALGenericIteratorDetection(object):
         elif (self.loader._name == "TFRecordReaderClassification"):
             if (self.loader._one_hot_encoding == True):
                 if self.device == "cpu":
-                    self.labels = np.zeros((self.bs) * (self.loader._num_classes), dtype="int32")
-                    self.loader.get_one_hot_encoded_labels(self.labels, device="cpu")
-                    self.labels = np.reshape(self.labels, (-1, self.bs, self.loader._num_classes))
+                    self.labels = np.zeros(
+                        (self.bs) * (self.loader._num_classes), dtype="int32")
+                    self.loader.get_one_hot_encoded_labels(
+                        self.labels, device="cpu")
+                    self.labels = np.reshape(
+                        self.labels, (-1, self.bs, self.loader._num_classes))
                 else:
-                    self.labels = cp.zeros((self.bs) * (self.loader._num_classes), dtype="int32")
-                    self.loader.get_one_hot_encoded_labels(self.labels, device="gpu")
-                    self.labels = cp.reshape(self.labels, (-1, self.bs, self.loader._num_classes))
+                    self.labels = cp.zeros(
+                        (self.bs) * (self.loader._num_classes), dtype="int32")
+                    self.loader.get_one_hot_encoded_labels(
+                        self.labels, device="gpu")
+                    self.labels = cp.reshape(
+                        self.labels, (-1, self.bs, self.loader._num_classes))
             else:
                 self.labels = self.loader.get_image_labels()
 
@@ -170,6 +186,7 @@ class ROCALGenericIteratorDetection(object):
     def __del__(self):
         b.rocalRelease(self.loader._handle)
 
+
 class ROCALIterator(ROCALGenericIteratorDetection):
     """
     ROCAL iterator for detection and classification tasks for PyTorch. It returns 2 or 3 outputs
@@ -183,6 +200,7 @@ class ROCALIterator(ROCALGenericIteratorDetection):
 
 
     """
+
     def __init__(self,
                  pipelines,
                  size=0,
@@ -194,7 +212,8 @@ class ROCALIterator(ROCALGenericIteratorDetection):
                  device_id=0):
         pipe = pipelines
         super(ROCALIterator, self).__init__(pipe, tensor_layout=pipe._tensor_layout, tensor_dtype=pipe._tensor_dtype,
-                                                            multiplier=pipe._multiplier, offset=pipe._offset, device=device, device_id=device_id)
+                                            multiplier=pipe._multiplier, offset=pipe._offset, device=device, device_id=device_id)
+
 
 class ROCAL_iterator(ROCALGenericImageIterator):
     """
